add some path routes and change the bruno folders

+++ course-ms/src/main/java/org/cognitia/course_ms/domain/path/dto/PathWithVideoDto.java
@@ -2,7 +2,7 @@ package org.cognitia.course_ms.domain.path.dto;
 
 public record PathWithVideoDto(
         String videoTitle,
-        Long videoDuration
+        Double videoDuration
 
 ) {
 }
